made proper package
